core/requirements.txt
opencv-python>=4.9.0
numpy>=1.24.0
pyautogui>=0.9.54
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

